+++ R/filter_missing.R
@@ -118,9 +118,7 @@ prep.step_filter_missing <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_filter_missing <- function(object, new_data, ...) {
-  if (length(object$removals) > 0) {
-    new_data <- new_data[, !(colnames(new_data) %in% object$removals)]
-  }
+  new_data <- recipes_remove_cols(new_data, object)
   new_data
 }