+++ R/ns.R
@@ -63,7 +63,6 @@ step_ns <-
            objects = NULL,
            deg_free = 2,
            options = list(),
-           keep_original_cols = FALSE,
            skip = FALSE,
            id = rand_id("ns")) {
     add_step(
@@ -75,7 +74,6 @@ step_ns <-
         role = role,
         objects = objects,
         options = options,
-        keep_original_cols = keep_original_cols,
         skip = skip,
         id = id
       )
@@ -83,8 +81,7 @@ step_ns <-
   }
 
 step_ns_new <-
-  function(terms, role, trained, deg_free, objects, options, keep_original_cols,
-           skip, id) {
+  function(terms, role, trained, deg_free, objects, options, skip, id) {
     step(
       subclass = "ns",
       terms = terms,
@@ -93,7 +90,6 @@ step_ns_new <-
       deg_free = deg_free,
       objects = objects,
       options = options,
-      keep_original_cols = keep_original_cols,
       skip = skip,
       id = id
     )
@@ -152,7 +148,6 @@ prep.step_ns <- function(x, training, info = NULL, ...) {
     deg_free = x$deg_free,
     objects = obj,
     options = x$options,
-    keep_original_cols = get_keep_original_cols(x),
     skip = x$skip,
     id = x$id
   )
@@ -176,7 +171,7 @@ bake.step_ns <- function(object, new_data, ...) {
       paste(orig_var, "ns", names0(new_cols[i], ""), sep = "_")
     colnames(ns_values)[cols] <- new_names
     strt <- max(cols) + 1
-    new_data <- remove_original_cols(new_data, object, orig_var)
+    new_data[, orig_var] <- NULL
   }
   ns_values <- as_tibble(ns_values)
   ns_values <- check_name(ns_values, new_data, object, names(ns_values))