+++ R/nzv.R
@@ -167,9 +167,7 @@ prep.step_nzv <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_nzv <- function(object, new_data, ...) {
-  if (length(object$removals) > 0) {
-    new_data <- new_data[, !(colnames(new_data) %in% object$removals)]
-  }
+  new_data <- recipes_remove_cols(new_data, object)
   new_data
 }