+++ NEWS.md
@@ -16,7 +16,7 @@
 
 * Added developer function `remove_original_cols()` to help remove original columns that are no longer needed.
 
-* `step_bs()` and `step_ns()` have gained `keep_original_cols` argument.
+* Added developer function `recipes_remove_cols()` to provide standardized way to remove columns by column names. (#1155)
 
 # recipes 1.0.6